Moved `ControllerHandle` to the `controller` module
Previously the `ControllerHandle` was defined in the `types` module.
It is much more clear for it to be part of the `controller` module.
Additionally, the `use` statements have been updated to reflect this.
