+++ src/types.rs
@@ -1,11 +1,7 @@
 //! Collection of types to increase readability and maintainability of the
 //! crate.
 
-use std::{
-    any::Any,
-    sync::mpsc::Sender,
-    thread::{JoinHandle, Thread},
-};
+use std::{any::Any, sync::mpsc::Sender};
 
 /// Shallow wrapper for a trait object using `Box` that can pass through thread
 /// boundaries.
@@ -68,44 +64,6 @@ pub enum JoinPattern {
     TernaryRecv(crate::patterns::ternary::RecvJoinPattern),
     /// Two `SendChannel` and `BidirChannel` Join Pattern.
     TernaryBidir(crate::patterns::ternary::BidirJoinPattern),
-}
-
-/// Handle to a `Junction`'s underlying `Controller`.
-///
-/// This struct carries a `JoinHandle` to the thread that the `Controller` of
-/// a `Junction` is running in. It allows for the `Controller` and its thread
-/// to be stopped gracefully at any point.
-pub struct ControllerHandle {
-    sender: Sender<Packet>,
-    control_thread_handle: Option<JoinHandle<()>>,
-}
-
-impl ControllerHandle {
-    pub(crate) fn new(sender: Sender<Packet>, handle: JoinHandle<()>) -> ControllerHandle {
-        ControllerHandle {
-            sender,
-            control_thread_handle: Some(handle),
-        }
-    }
-
-    /// Extracts a handle to the underlying thread.
-    pub fn thread(&self) -> Option<&Thread> {
-        match &self.control_thread_handle {
-            Some(h) => Some(h.thread()),
-            None => None,
-        }
-    }
-
-    /// Request the `Controller` to stop gracefully, then join its thread.
-    ///
-    /// # Panics
-    ///
-    /// Panics if it was unable to send shut-down request to the control thread.
-    pub fn stop(&mut self) {
-        self.sender.send(Packet::ShutDownRequest).unwrap();
-
-        self.control_thread_handle.take().unwrap().join().unwrap();
-    }
 }
 
 /// Function types related to various kind of functions that can be stored and
